+++ tests/os_log_entry_activity_tests.rs
@@ -15,7 +15,7 @@ fn os_log_entry_activity_accessors() {
 
     let store = OSLogStore::new(OSLogStoreScope::CurrentProcessIdentifier).expect("store");
     let entries = store
-        .entries(
+        .get_entries(
             OSLogEnumeratorOptions::REVERSE,
             Some(&store.position_time_interval_since_end(Duration::from_secs(5))),
             &OSLogEntryFilter::default(),