+++ tests/os_log_entry_log_tests.rs
@@ -8,7 +8,7 @@ fn os_log_entry_log_accessors() {
     std::thread::sleep(Duration::from_millis(100));
     let store = OSLogStore::new(OSLogStoreScope::CurrentProcessIdentifier).expect("store");
     let entries = store
-        .entries(
+        .get_entries(
             OSLogEnumeratorOptions::REVERSE,
             Some(&store.position_time_interval_since_end(Duration::from_secs(5))),
             &OSLogEntryFilter::default(),